rego: split off some code
